fix: определить интерфейс UserService существующий класс UserService переименовать в UserServiceImpl, реализующим интерфейс UserService в классе-контроллере UserController заменить класс на интерфейс
